requirements.txt
fastapi
uvicorn
Pillow
requests